+++ R/azcopy.R
@@ -1,7 +1,7 @@
 #' Call the azcopy file transfer utility
 #'
+#' @param object An AzureStor object from which to obtain authentication details. Can be either a storage endpoint or container.
 #' @param ... Arguments to pass to AzCopy on the commandline. If no arguments are supplied, a help screen is printed.
-#' @param env Environment variables to pass to AzCopy. Typically these will be for authentication.
 #'
 #' @details
 #' AzureStor has the ability to use the Microsoft AzCopy commandline utility to transfer files. To enable this, set the argument `use_azcopy=TRUE` in any call to an upload or download function; AzureStor will then call AzCopy to perform the file transfer rather than relying on its own code. You can also call AzCopy directly with the `call_azcopy` function, passing it any arguments as required.
@@ -10,81 +10,82 @@
 #'
 #' The AzCopy utility must be in your path for AzureStor to find it. Note that unlike earlier versions, Azcopy 10 is a single, self-contained binary file that can be placed in any directory.
 #'
-#' AzCopy uses its own mechanisms for authenticating with Azure Active Directory, which is independent of the OAuth tokens used by AzureStor. AzureStor will try to ensure that AzCopy has previously authenticated before trying to transfer a file with a token, but this may not always succeed. You can run `azcopy_login(force=TRUE)` to force it to authenticate.
-#'
+#' @return
+#' A list, invisibly, with the following components:
+#' - `status`: The exit status of the AzCopy command. If this is NA, then the process was killed and had no exit status.
+#' - `stdout`: The standard output of the command.
+#' - `stderr`: The standard error of the command.
+#' - `timeout`: Whether AzCopy was killed because of a timeout.
 #' @seealso
+#' [processx::run], [download_blob], [download_azure_file], [download_adls_file]
+#'
 #' [AzCopy page on Microsoft Docs](https://docs.microsoft.com/en-us/azure/storage/common/storage-use-azcopy-v10)
 #'
 #' [AzCopy GitHub repo](https://github.com/Azure/azure-storage-azcopy)
+#' @examples
+#' \dontrun{
+#'
+#' endp <- storage_endpoint("https://mystorage.blob.core.windows.net", sas="mysas")
+#' cont <- storage_container(endp, "mycontainer")
+#'
+#' # print various help screens
+#' call_azcopy(endp, "help")
+#' call_azcopy(endp, "help", "copy")
+#' call_azcopy(endp, "help", "sync")
+#'
+#' # calling azcopy to download a blob
+#' storage_download(cont, "myblob.csv", use_azcopy=TRUE)
+#'
+#' # calling azcopy directly (must specify the SAS explicitly in the source URL)
+#' call_azcopy(endp,
+#'             "copy",
+#'             "https://mystorage.blob.core.windows.net/mycontainer/myblob.csv?mysas",
+#'             "myblob.csv")
+#'
+#' }
 #' @aliases azcopy
 #' @rdname azcopy
 #' @export
-call_azcopy <- function(..., env=NULL)
+call_azcopy <- function(object, ...)
+{
+    UseMethod("call_azcopy")
+}
+
+#' @export
+call_azcopy.storage_container <- function(object, ...)
+{
+    call_azcopy(object$endpoint, ...)
+}
+
+#' @export
+call_azcopy.storage_endpoint <- function(object, ...)
 {
     if(!requireNamespace("processx"))
         stop("The processx package must be installed to use azcopy", call.=FALSE)
 
-    azcopy <- get_azcopy_path()
+    auth <- azcopy_auth(object)
     args <- as.character(unlist(list(...)))
-    invisible(processx::run(azcopy, args, echo=TRUE, echo_cmd=TRUE, env=env))
+    call_azcopy_internal(args, auth$env)
 }
 
 
-# azcopy unset/NULL -> not initialized
-# azcopy = NA -> binary not found, or version < 10 (not usable)
-# azcopy = path -> usable
-get_azcopy_path <- function()
+call_azcopy_internal <- function(args, env)
 {
-    if(exists("azcopy", envir=.AzureStor))
-    {
-        if(!is.na(.AzureStor$azcopy))
-            return(.AzureStor$azcopy)
-        else stop("azcopy version 10+ required but not found", call.=FALSE)
-    }
-    else
-    {
-        set_azcopy_path()
-        Recall()
-    }
-}
-
-
-set_azcopy_path <- function(path="azcopy")
-{
-    path <- Sys.which(path)
-    if(is.na(path) || path == "")
-    {
-        .AzureStor$azcopy <- NA
-        return(NULL)
-    }
-
-    ver <- suppressWarnings(processx::run(path, "--version"))
-    if(!grepl("version 1[[:digit:]]", ver$stdout, ignore.case=TRUE))
-    {
-        .AzureStor$azcopy <- NA
-        return(NULL)
-    }
-
-    .AzureStor$azcopy <- unname(path)
-    message("Using azcopy binary ", path)
-    invisible(NULL)
+    print(env)
+    invisible(processx::run(get_azcopy_path(), args, echo_cmd=TRUE, echo=TRUE, env=env))
 }
 
 
 azcopy_upload <- function(container, src, dest, ...)
 {
     opts <- azcopy_upload_opts(container, ...)
+
     dest_uri <- httr::parse_url(container$endpoint$url)
     dest_uri$path <- gsub("//", "/", file.path(container$name, dest))
-    dest <- httr::build_url(dest_uri)
+    dest <- azcopy_add_sas(container$endpoint, httr::build_url(dest_uri))
 
-    auth <- azcopy_auth(container)
-    azcopy_login(auth)
-    dest <- azcopy_add_sas(auth, dest)
-
-    call_azcopy("copy", src, dest, opts, env=auth$env)
+    call_azcopy(container$endpoint, "copy", src, dest, opts)
 }
-
 
 azcopy_upload_opts <- function(container, ...)
 {
@@ -111,17 +112,13 @@ azcopy_upload_opts.adls_filesystem <- function(container, blocksize=2^24, recurs
 azcopy_download <- function(container, src, dest, ...)
 {
     opts <- azcopy_download_opts(container, ...)
+
     src_uri <- httr::parse_url(container$endpoint$url)
     src_uri$path <- gsub("//", "/", file.path(container$name, src))
-    src <- httr::build_url(src_uri)
+    src <- azcopy_add_sas(container$endpoint, httr::build_url(src_uri))
 
-    auth <- azcopy_auth(container)
-    azcopy_login(auth)
-    src <- azcopy_add_sas(auth, src)
-
-    call_azcopy("copy", src, dest, opts, env=auth$env)
+    call_azcopy("copy", src, dest, opts)
 }
-
 
 azcopy_download_opts <- function(container, ...)
 {